[PaymentRequest] Move PaymentRequestHashAware interface and trait to PaymentBundle
