+++ src/Sylius/Bundle/PaymentBundle/Command/PaymentRequestHashAwareInterface.php
@@ -11,9 +11,8 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Bundle\CoreBundle\PaymentRequest\Command;
+namespace Sylius\Bundle\PaymentBundle\Command;
 
-/** @experimental */
 interface PaymentRequestHashAwareInterface
 {
     public function getHash(): string;
