+++ src/Sylius/Bundle/PaymentBundle/Command/PaymentRequestHashAwareTrait.php
@@ -11,7 +11,7 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Bundle\CoreBundle\PaymentRequest\Command;
+namespace Sylius\Bundle\PaymentBundle\Command;
 
 trait PaymentRequestHashAwareTrait
 {
